Route fixes for auth

+++ routes/web.php
@@ -18,16 +18,16 @@ use App\Http\Controllers\ListingController;
 
 // Listings
 Route::get('/', [ListingController::class, 'index']);
-Route::get('/listings/create', [ListingController::class, 'create']);
-Route::post('/listings', [ListingController::class, 'store']);
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 // Users
-Route::get('/users/register', [UserController::class, 'create']);
+Route::get('/users/register', [UserController::class, 'create'])->middleware('guest');
 Route::post('/users', [UserController::class, 'store']);
-Route::post('/users/logout', [UserController::class, 'logout']);
-Route::get('/users/login', [UserController::class, 'login']);
+Route::post('/users/logout', [UserController::class, 'logout'])->middleware('auth');
+Route::get('/users/login', [UserController::class, 'login'])->name('login')->middleware('guest');;
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
